refactor: Reorganizar lógica de autenticación y manejo de usuarios en authController
